same functionality, only with javascript
