+++ package.json
@@ -3,33 +3,39 @@
   "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@types/react-router-dom": "^4.3.0",
     "color-rainbow": "0.0.0",
     "elasticsearch": "^15.1.1",
     "leaflet": "^1.3.4",
-    "react": "^16.8.6",
+    "react": "^16.9.0",
     "react-debounce-input": "^3.2.0",
-    "react-dom": "^16.8.6",
+    "react-dom": "^16.9.0",
     "react-leaflet": "^2.0.1",
     "react-router-dom": "^4.3.1",
-    "react-scripts-ts": "^3.1.0"
-  },
-  "scripts": {
-    "start": "PORT=8080 react-scripts-ts start",
-    "build": "react-scripts-ts build && cp _redirects build/_redirects",
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject"
+    "react-scripts": "3.1.1"
   },
   "devDependencies": {
-    "@types/elasticsearch": "^5.0.26",
-    "@types/geojson": "^7946.0.4",
-    "@types/leaflet": "^1.2.11",
-    "@types/node": "^10.11.3",
-    "@types/react": "^16.4.14",
-    "@types/react-dom": "^16.0.7",
-    "@types/react-leaflet": "^1.1.6",
     "leaflet.awesome-markers": "^2.0.4",
-    "prop-types": "^15.6.2",
-    "typescript": "^3.1.1"
+    "prop-types": "^15.7.2"
+  },
+  "scripts": {
+    "start": "PORT=8080 react-scripts start",
+    "build": "react-scripts build && cp _redirects build/_redirects",
+    "test": "react-scripts test --env=jsdom",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
