+++ src/Components/SearchResults.jsx
@@ -1,0 +1,59 @@
+import * as React from 'react';
+import { MeinItems } from './SearchResults/MeinItem';
+import SearchResultDetailled from './SearchResultDetailled';
+import './SearchResults.css';
+/**
+ * This component contains the various search results, displaying one SearchResults-component (not this one, but the one in sub folder SearchResults) for each one.
+ * If a result was selected, display a SearchResultDetailled for that result.
+ */
+class SearchResults extends React.Component {
+    constructor() {
+        super(...arguments);
+        /**
+         * Onclick handler for search result row, updates selectedId in searchParams
+         */
+        this.toggleSelection = (event, id) => {
+            let searchParams = this.props.searchParams;
+            if (searchParams.selectedId === id) {
+                searchParams.selectedId = 0;
+            }
+            else {
+                searchParams.selectedId = id;
+            }
+            this.props.updateHandler(searchParams);
+        };
+    }
+    render() {
+        let results = this.props.results || [];
+        const searchParams = this.props.searchParams;
+        if (searchParams.selectedId) {
+            const result = results.find((element) => searchParams.selectedId === element.id);
+            if (result) {
+                return (<SearchResultDetailled result={result} searchParams={this.props.searchParams} updateHandler={this.props.updateHandler}/>);
+            }
+        }
+        results = results.filter(element => {
+            let date = null;
+            if (element.dateStart !== null) {
+                date = new Date(element.dateStart);
+            }
+            return element.type !== 'event' || (date === null || date.getDay() === new Date().getDay());
+        });
+        // limit the number of displayed results. If none was given display the first 200
+        const limit = this.props.limit ? this.props.limit : 200;
+        if (results && results.length > limit) {
+            results = results.slice(0, this.props.limit);
+        }
+        return (<div className="search_results">
+        {results.map((result) => {
+            const meinItem = MeinItems.getItem(result.type);
+            const ComponentClass = meinItem.component;
+            const searchResultComponent = <ComponentClass result={result} icon={meinItem.icon}/>;
+            return (<article key={result.id} className={'notification' + ((searchParams.selectedId === result.id) ? ' ' + meinItem.color : '')} onClick={e => this.toggleSelection(e, result.id)}>
+              {searchResultComponent}
+            </article>);
+        })}
+      </div>);
+    }
+}
+export default SearchResults;
