+++ src/Components/SearchResults/MeinItem.js
@@ -6,136 +6,126 @@ import SearchResultsWifi from './SearchResultsWifi';
 import SearchResultsWebcam from './SearchResultsWebcam';
 import SearchResultsDefault from './SearchResultsDefault';
 import SearchResultsEvent from './SearchResultsEvent';
-
-export interface IMeinItem {
-    name: string;
-    icon: string;
-    color: string;
-    component: any;
-}
-
 /**
  * Class to hold colors, types, names and icons for all our Types
  */
-export abstract class MeinItems {
-
-    public static items: {[id: string]: IMeinItem} = {
-      babychange: {
+export class MeinItems {
+    static getItem(key) {
+        return MeinItems.items[key] || {
+            name: key,
+            icon: null,
+            color: 'is-light',
+            component: SearchResultsDefault
+        };
+    }
+}
+MeinItems.items = {
+    babychange: {
         name: 'Wickelraum',
         icon: 'mdi-baby-buggy',
         color: 'is-link',
         component: SearchResultsDefault
-      },
-      bookbus: {
+    },
+    bookbus: {
         name: 'Bücherbus',
         icon: 'mdi-bus-clock',
         color: 'is-link',
         component: SearchResultsDefault
-      },
-      cemetery: {
+    },
+    cemetery: {
         name: 'Friedhof',
         icon: 'mdi-grave-stone',
         color: 'is-light',
         component: SearchResultsDefault
-      },
-      container: {
+    },
+    container: {
         name: 'Recycling-Container',
         icon: 'mdi-recycle',
         color: 'is-danger',
         component: SearchResultsDefault
-      },      
-      construction: {
+    },
+    construction: {
         name: 'Baustelle',
         icon: 'mdi-vlc',
         color: 'is-danger',
         component: SearchResultsConstruction
-      },
-      kindergarden: {
+    },
+    kindergarden: {
         name: 'Kindergarten',
         icon: 'mdi-baby',
         color: 'is-primary',
         component: SearchResultsDefault
-      },
-      lunch: {
+    },
+    lunch: {
         name: 'Mittagstisch',
         icon: 'mdi-food',
         color: 'is-link',
         component: SearchResultsLunch
-      },
-      pool: {
+    },
+    pool: {
         name: 'Bad',
         icon: 'mdi-pool',
         color: 'is-info',
         component: SearchResultsDefault
-      },
-      playground: {
+    },
+    playground: {
         name: 'Spielplatz',
         icon: 'mdi-castle',
         color: 'is-warning',
         component: SearchResultsPlayground
-      },
-      wc: {
+    },
+    wc: {
         name: 'Toilette',
         icon: 'mdi-human-male-female',
         color: 'is-light',
         component: SearchResultsWc
-      },
-      wifi: {
+    },
+    wifi: {
         name: 'WLAN',
         icon: 'mdi-wifi',
         color: 'is-dark',
         component: SearchResultsWifi
-      },
-      webcam: {
+    },
+    webcam: {
         name: 'Webcam',
         icon: 'mdi-camera',
         color: 'is-success',
         component: SearchResultsWebcam
-      },
-      event: {
+    },
+    event: {
         name: 'Termin',
         icon: 'mdi-calendar-text',
         color: 'is-black',
         component: SearchResultsEvent
-      },
-      school: {
+    },
+    school: {
         name: 'Schule',
         icon: 'mdi-school',
         color: 'is-light',
         component: SearchResultsDefault
-      },
-      sport: {
+    },
+    sport: {
         name: 'Sportstätte',
         icon: 'mdi-soccer',
         color: 'is-success',
         component: SearchResultsDefault
-      },
-      agencies: {
+    },
+    agencies: {
         name: 'Amt',
         icon: 'mdi-briefcase',
         color: 'is-black',
         component: SearchResultsDefault
-      },
-      papierkorb: {
+    },
+    papierkorb: {
         name: 'Papierkorb',
         icon: 'mdi-trash-can-outline',
         color: 'is-black',
         component: SearchResultsDefault
-      },
-      library: {
+    },
+    library: {
         name: 'Bücherei',
         icon: 'mdi-book',
         color: 'is-warning',
         component: SearchResultsDefault
-      }
-    };
-
-    public static getItem(key: string): IMeinItem {
-      return MeinItems.items[key] || {
-        name: key,
-        icon: null,
-        color: 'is-light',
-        component: SearchResultsDefault
-      };
     }
-}
+};
