+++ src/Components/SearchResults/SearchResultsPlayground.jsx
@@ -1,17 +1,13 @@
 import * as React from 'react';
 import { SearchResultsBase } from './SearchResultsBase';
-
 class SearchResultsPlayground extends SearchResultsBase {
-
-  render() {
-    const result = this.props.result;
-
-    return (
-      <div className="media">
+    render() {
+        const result = this.props.result;
+        return (<div className="media">
         <div className="media-left">
           <p>
             <span className="icon is-large">
-              <i className="mdi mdi-48px mdi-castle" />
+              <i className="mdi mdi-48px mdi-castle"/>
             </span>
           </p>
           <div className="distanceDiv has-text-centered">
@@ -25,15 +21,10 @@ class SearchResultsPlayground extends SearchResultsBase {
               <span className="tag is-dark">Spielplatz</span> &nbsp;
               {result.properties.Groesse && <span className="tag is-dark"> {result.properties.Groesse} m<sup>2</sup></span>} &nbsp;
               {this.displayPitch(result.properties.Ball, result.properties.Streetball)
-                &&
-                <span
-                  data-balloon="Mit Ballspielplatz"
-                  data-balloon-pos="up"
-                  className="icon"
-                >
-                    <i className="mdi mdi-soccer" />
-                </span>
-              } &nbsp;
+            &&
+                <span data-balloon="Mit Ballspielplatz" data-balloon-pos="up" className="icon">
+                    <i className="mdi mdi-soccer"/>
+                </span>} &nbsp;
               {this.displaySkater(result.properties.Skater) && <span> <img className="tag-image" src="/media/skateboard.png"/> </span>} &nbsp;
               {result.properties.Bereich && <span className="tag is-success"> {this.displayArea(result.properties.Bereich)} </span>} &nbsp;
             </span>
@@ -42,59 +33,50 @@ class SearchResultsPlayground extends SearchResultsBase {
             </div>
             <p className="has-text-danger">
               <span className="icon">
-                <i className="mdi mdi-walk" />
+                <i className="mdi mdi-walk"/>
               </span>
               {this.getMinutesByFeet(result.distance)}
               &bull;
               <span className="icon">
-                <i className="mdi mdi-car" />
+                <i className="mdi mdi-car"/>
               </span>
               {this.getMinutesByCar(result.distance)}
             </p>
           </div>
         </div>
-      </div>
-    );
-  }
-
-  private displayName(name: string): string {
-    let result = '';
-    result = name.slice(3);
-
-    return result.toLowerCase().replace(/\b\w/g, function(txt: any) {
-      return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
-    });
-  }
-
-  private displayArea(area: string): string {
-    let result = '';
-
-    switch (area) {
-    case 'A':
-        result = 'Großer Spielplatz';
-        break;
-    case 'B/C':
-        result = 'Wohngebiets-Spielplatz';
-        break;
-    case 'C':
-        result = 'Für Kleinkinder';
-        break;
-    case 'A+B/C':
-        result = 'Groß, und extra Kleinkinderbereich';
-        break;
-    default:
+      </div>);
     }
-
-    return result;
-  }
-
-  private displayPitch(ball: number, streetBall: string): boolean {
-    return ball === 1 || ball === 2 || streetBall === 'ja';
-  }
-
-  private displaySkater(skater: string): boolean {
-    return skater === 'ja';
-  }
+    displayName(name) {
+        let result = '';
+        result = name.slice(3);
+        return result.toLowerCase().replace(/\b\w/g, function (txt) {
+            return txt.charAt(0).toUpperCase() + txt.substr(1).toLowerCase();
+        });
+    }
+    displayArea(area) {
+        let result = '';
+        switch (area) {
+            case 'A':
+                result = 'Großer Spielplatz';
+                break;
+            case 'B/C':
+                result = 'Wohngebiets-Spielplatz';
+                break;
+            case 'C':
+                result = 'Für Kleinkinder';
+                break;
+            case 'A+B/C':
+                result = 'Groß, und extra Kleinkinderbereich';
+                break;
+            default:
+        }
+        return result;
+    }
+    displayPitch(ball, streetBall) {
+        return ball === 1 || ball === 2 || streetBall === 'ja';
+    }
+    displaySkater(skater) {
+        return skater === 'ja';
+    }
 }
-
 export default SearchResultsPlayground;
