--- tsconfig.prod.json
@@ -1,1 +1,0 @@
-tsconfig.json
